Implement role management module with CRUD operations, including UI components for adding, editing, and deleting roles. Integrate API handling and state management using Redux. Update routing and sidebar for roles access.

+++ src/hooks/useAuthCheck.js
@@ -1,14 +1,17 @@
-// hooks/useAuthCheck.js
 import { useEffect, useState } from "react";
 import { useDispatch } from "react-redux";
 import Cookies from "js-cookie";
-import { checkAuthAsync } from "@/stores/slices/authSlice";
+import {
+  checkAuthAsync,
+  selectFetchingStatus,
+} from "@/stores/slices/authSlice";
 import { verifyJwtToken } from "@/utils/helper";
 import { store } from "@/stores";
+import { useSelector } from "react-redux";
 
 export const useAuthCheck = () => {
   const dispatch = useDispatch();
-  const [status, setStatus] = useState({ loading: true, isValid: false });
+  const status = useSelector(selectFetchingStatus);
 
   useEffect(() => {
     const token = localStorage.getItem("accessToken");
@@ -17,25 +20,19 @@ export const useAuthCheck = () => {
     if (!isTokenValid) {
       Cookies.remove("accessToken");
       localStorage.clear();
-      setStatus({ loading: false, isValid: false });
       return;
     }
 
+    if (status === "loading") return;
     dispatch(checkAuthAsync())
       .unwrap()
-      .then((res) => {
-        setStatus({
-          loading: false,
-          isValid: !!res?.data?.user,
-        });
-      })
+      .then(() => {})
       .catch(() => {
         Cookies.remove("accessToken");
         localStorage.clear();
         store.dispatch({ type: "RESET_STORE" });
-        setStatus({ loading: false, isValid: false });
       });
   }, []);
 
-  return status;
+  return { loading: status === "loading", isValid: status === "idle" };
 };
